chore/ 메인경로 수정 및 인플루언서 폴더 통일
