+++ src/app/(advertiser)/components/about/AboutSection.tsx
