+++ src/app/(advertiser)/components/about/Slide.tsx
