+++ src/app/(advertiser)/components/about/Slides.tsx
