+++ src/app/(advertiser)/components/intro/IntroSection.tsx
