+++ src/app/(advertiser)/components/logoline/LogoLine.tsx
