+++ src/app/(advertiser)/components/reference/ReferenceSlider.tsx
