+++ src/app/(advertiser)/components/reference/slideData.ts
