--- src/app/(influencer)/page.tsx
@@ -1,21 +1,0 @@
-import MainSection from './_view/mainSection';
-import Product from './_view/product';
-import Introduction from './_view/introduction';
-import Banner from './_view/banner';
-import Review from './_view/review';
-import Question from './_view/question';
-import Join from './_view/join';
-
-export default function InfluencerPage() {
-  return (
-    <>
-      <MainSection />
-      <Product />
-      <Introduction />
-      <Banner />
-      <Review />
-      <Question />
-      <Join />
-    </>
-  );
-}
