+++ src/app/influencer/_view/mainSection.tsx
@@ -4,30 +4,31 @@ import goolePlayQr from '/public/assets/images/qr_playstore.png';
 import Image from 'next/image';
 
 export default function MainSection() {
+  const videoId = '8bBoGxyY48E';
   return (
     <div className='mt-[58px] w-full bg-cardMint'>
       <section
-        className='flex aspect-[404/227] max-h-[404px] w-full justify-center'
+        className='flex w-full justify-center xs:aspect-[404/227] xs:max-h-[404px]'
         style={{
           backgroundImage: `url(${backgroundBubble.src})`,
           backgroundSize: 'contain',
           backgroundRepeat: 'repeat-x',
         }}
       >
-        <article className='m-auto mx-auto flex aspect-[720/227] w-full max-w-[720px] items-center justify-between'>
-          <div className='flex flex-col gap-y-[50px] px-[50px]'>
-            <div className='flex flex-col gap-5 text-mainText'>
-              <h2 className='text-[clamp(12px,2.5vw,1rem)] font-medium tracking-[-0.8px]'>
+        <article className='m-auto mx-auto flex w-full max-w-[720px] flex-col items-center justify-between xs:aspect-[720/227] xs:flex-row'>
+          <div className='flex flex-col gap-y-[50px] pt-[44px] xs:px-[50px]'>
+            <div className='flex flex-col gap-y-3 text-center text-mainText xs:gap-5 xs:text-left'>
+              <h2 className='font-medium tracking-[-0.8px] xs:text-[clamp(12px,2.5vw,1rem)]'>
                 인플루언서 마케팅 NO.1
               </h2>
 
-              <h1 className='text-[clamp(16px,4vw,2rem)] font-bold leading-tight'>
+              <h1 className='text-2xl font-bold leading-tight xs:text-[clamp(16px,4vw,2rem)]'>
                 여행 인플루언서
                 <br />
                 마케팅의 시작, 와이리
               </h1>
             </div>
-            <ul className='flex origin-top-left scale-[min(1,max(0.5,100vw/1440))] gap-3'>
+            <ul className='hidden origin-top-left scale-[min(1,max(0.5,100vw/1440))] gap-3 xs:flex'>
               <li className='flex items-center justify-center rounded-[16px] bg-white p-[14px] shadow-qrBoxShadow'>
                 <div className='flex w-[80px] flex-col gap-y-2'>
                   <Image
@@ -95,14 +96,90 @@ export default function MainSection() {
                 </div>
               </li>
             </ul>
+            <div className='mb-[66px] flex gap-[10px] xs:hidden'>
+              <button
+                type='button'
+                className='flex h-[42px] items-center justify-center gap-x-[8px] rounded-[56px] bg-primaryColor px-[24px] font-semibold tracking-[-5%] text-[#fff] hover:bg-[#3ECCCC] hover:shadow-qrBoxShadow'
+              >
+                <svg
+                  className='h-[18px] w-[18px]'
+                  width='31'
+                  height='30'
+                  viewBox='0 0 31 30'
+                  fill='none'
+                  xmlns='http://www.w3.org/2000/svg'
+                >
+                  <g clipPath='url(#clip0_177_3539)'>
+                    <path
+                      d='M25.7441 9.429C25.6211 9.506 22.6931 11.131 22.6931 14.734C22.8311 18.843 26.3881 20.284 26.4491 20.284C26.3881 20.361 25.9121 22.247 24.5021 24.224C23.3831 25.927 22.1411 27.644 20.2551 27.644C18.4611 27.644 17.8171 26.509 15.7471 26.509C13.5241 26.509 12.8951 27.644 11.1931 27.644C9.30712 27.644 7.97312 25.835 6.79312 24.148C5.26012 21.94 3.95712 18.475 3.91112 15.148C3.88012 13.385 4.21812 11.652 5.07612 10.18C6.28712 8.125 8.44912 6.73 10.8101 6.684C12.6191 6.623 14.2291 7.926 15.3331 7.926C16.3911 7.926 18.3691 6.684 20.6071 6.684C21.5731 6.685 24.1491 6.976 25.7441 9.429ZM15.1801 6.332C14.8581 4.722 15.7471 3.112 16.5751 2.085C17.6331 0.843 19.3041 0 20.7451 0C20.8371 1.61 20.2541 3.189 19.2121 4.339C18.2771 5.581 16.6671 6.516 15.1801 6.332Z'
+                      fill='#FFF'
+                    />
+                  </g>
+                  <defs>
+                    <clipPath id='clip0_177_3539'>
+                      <rect
+                        width='30'
+                        height='30'
+                        fill='white'
+                        transform='translate(0.5)'
+                      />
+                    </clipPath>
+                  </defs>
+                </svg>
+                APP Store
+              </button>
+              <button
+                type='button'
+                className='flex h-[42px] items-center justify-center gap-x-[8px] rounded-[56px] bg-primaryColor px-[24px] font-semibold tracking-[-5%] text-[#fff] hover:bg-[#3ECCCC] hover:shadow-qrBoxShadow'
+              >
+                <svg
+                  className='h-[18px] w-[18px]'
+                  width='31'
+                  height='30'
+                  viewBox='0 0 31 30'
+                  fill='none'
+                  xmlns='http://www.w3.org/2000/svg'
+                >
+                  <path
+                    d='M16.675 15.0008L5.06875 26.607C4.9375 26.332 4.875 26.032 4.875 25.7195V4.28203C4.875 3.96953 4.9375 3.66953 5.06875 3.39453L16.675 15.0008Z'
+                    fill='white'
+                  />
+                  <path
+                    d='M18.0004 16.3252L21.8816 20.2064L8.10039 27.6439C7.77539 27.8189 7.41914 27.9064 7.06289 27.9064C6.86914 27.9064 6.68164 27.8814 6.49414 27.8314L18.0004 16.3252Z'
+                    fill='white'
+                  />
+                  <path
+                    d='M18 13.6754L6.5 2.17539C7.03125 2.02539 7.6 2.08789 8.1 2.35664L21.8812 9.79414L18 13.6754Z'
+                    fill='white'
+                  />
+                  <path
+                    d='M29.1189 15C29.1189 15.8062 28.6752 16.5437 27.9689 16.925L23.6064 19.2812L19.3252 15L23.6064 10.7188L27.9689 13.075C28.6752 13.4562 29.1189 14.1937 29.1189 15Z'
+                    fill='white'
+                  />
+                </svg>
+                Google Play
+              </button>
+            </div>
           </div>
-          <div
-            className='aspect-[227/404] w-full max-w-[227px] bg-cardYellow'
-            style={{
-              transform: `scale(min(1, max(0.5, 100vw / 1440)))`,
-              transformOrigin: 'top left',
-            }}
-          ></div>
+          <div className='w-full xs:max-w-[227px]'>
+            <div
+              className='relative aspect-[227/404] w-full'
+              style={{
+                transform: `scale(min(1, max(0.5, 100vw / 1440)))`,
+                transformOrigin: 'top left',
+              }}
+            >
+              <iframe
+                className='absolute inset-0 h-full w-full'
+                src={`https://www.youtube.com/embed/${videoId}?autoplay=1&mute=1&controls=0&modestbranding=1&playsinline=1&rel=0&enablejsapi=1&widgetid=1&loop=1&playlist=${videoId}`}
+                title='YouTube Shorts player'
+                allow='accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture'
+                allowFullScreen
+              />
+              {/* 마우스 이벤트를 차단하는 오버레이 레이어 */}
+              <div className='pointer-events-auto absolute inset-0' />
+            </div>
+          </div>
         </article>
       </section>
     </div>
